+++ packages/ceramic-cli/src/ceramic-cli-utils.ts
@@ -6,7 +6,7 @@ const PREFIX_REGEX = /^ceramic:\/\/|^\/ceramic\//
 /**
  * Ceramic CLI utility functions
  */
-export default class CeramicCliUtils {
+export class CeramicCliUtils {
 
     /**
      * Create document
